+++ NEWS.md
@@ -1,3 +1,16 @@
+# Version 0.6.1
+* Fixed a bug for reading files with different overview levels.
+* Added additional early checks for grid shapes.
+
+# Version 0.6.0
+* Added `closed_border` argument to control whether cells outside the mask are excluded from the analysis.
+* Fixed the polygon-masked output to not return the buffered array.
+* Added compilation CPU optimisation flags (slower complie but faster run).
+* Changed the `overview_levels` argument of the `create_overviews()` function to `levels` for consistency. Also, `file_path` in `read_raster()` function.
+
+# Version 0.5.0
+* Added `has_overview()` function for checking overviews.
+
 # Version 0.4.2
 * A few internal structural changes; e.g., changes to the NumPy array memory layout and added structure in the Rust code for performance improvement.